requirements.txt
allure-pytest==2.8.12
allure-python-commons==2.8.12
attrs==21.4.0
autopep8==1.6.0
more-itertools==8.12.0
packaging==21.3
pluggy==0.13.1
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.7
pytest==5.4.1
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
wcwidth==0.2.5
